update facilities gis client and configuration namespacing (#3576)
