+++ lib/facilities/gis/metadata_client.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require 'common/client/base'
+
+module Facilities
+  module Gis
+    class MetadataClient < Facilities::MetadataClient
+      configuration Facilities::Gis::MetadataConfiguration
+    end
+  end
+end
